reorganize mocks and main
